remove curator
