+++ code/go/0chain.net/smartcontract/storagesc/allocation_test.go
@@ -535,103 +535,6 @@ func TestStorageSmartContract_getAllocation(t *testing.T) {
 	assert.Equal(t, alloc.Encode(), got.Encode())
 }
 
-func TestAddCurator(t *testing.T) {
-	t.Parallel()
-	const (
-		mockClientId     = "mock client id"
-		mockCuratorId    = "mock curator id"
-		mockAllocationId = "mock allocation id"
-	)
-	type args struct {
-		ssc      *StorageSmartContract
-		txn      *transaction.Transaction
-		input    []byte
-		balances chainState.StateContextI
-	}
-	type parameters struct {
-		clientId         string
-		info             addCuratorInput
-		existingCurators []string
-	}
-	type want struct {
-		err    bool
-		errMsg string
-	}
-	var setExpectations = func(t *testing.T, name string, p parameters, want want) args {
-		var balances = &mocks.StateContextI{}
-		var txn = &transaction.Transaction{
-			ClientID: p.clientId,
-		}
-		var ssc = &StorageSmartContract{
-			SmartContract: sci.NewSC(ADDRESS),
-		}
-		input, err := json.Marshal(p.info)
-		require.NoError(t, err)
-
-		var sa = StorageAllocation{
-			ID:    p.info.AllocationId,
-			Owner: p.clientId,
-		}
-		for _, curator := range p.existingCurators {
-			sa.Curators = append(sa.Curators, curator)
-		}
-		balances.On("GetTrieNode", sa.GetKey(ssc.ID)).Return(&sa, nil).Once()
-
-		balances.On(
-			"InsertTrieNode",
-			sa.GetKey(ssc.ID),
-			mock.MatchedBy(func(sa *StorageAllocation) bool {
-				if len(sa.Curators) < 1 {
-					return false
-				}
-				for i, curator := range p.existingCurators {
-					if curator != sa.Curators[i] {
-						return false
-					}
-				}
-				if p.info.CuratorId != sa.Curators[len(sa.Curators)-1] {
-					return false
-				}
-				return sa.ID == p.info.AllocationId && sa.Owner == p.clientId
-			})).Return("", nil).Once()
-
-		return args{ssc, txn, input, balances}
-	}
-
-	testCases := []struct {
-		name       string
-		parameters parameters
-		want       want
-	}{
-		{
-			name: "ok",
-			parameters: parameters{
-				clientId: mockClientId,
-				info: addCuratorInput{
-					CuratorId:    mockCuratorId,
-					AllocationId: mockAllocationId,
-				},
-			},
-		},
-	}
-	for _, test := range testCases {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-			args := setExpectations(t, test.name, test.parameters, test.want)
-
-			err := args.ssc.addCurator(args.txn, args.input, args.balances)
-
-			require.EqualValues(t, test.want.err, err != nil)
-			if err != nil {
-				require.EqualValues(t, test.want.errMsg, err.Error())
-				return
-			}
-			require.True(t, mock.AssertExpectationsForObjects(t, args.balances))
-		})
-	}
-}
-
 func TestTransferAllocation(t *testing.T) {
 	t.Parallel()
 	const (
